Patch 13.0 - Costs 1-6 with tiles in Dijkstra plus dijkstra goes fuckin' fast plus I'm happy plus looks cool! ^^ (plus random)

+++ src/Agent.h
@@ -47,7 +47,7 @@ public:
 
 	// Utils
 	vector<Vector2D> frontierCount;
-	vector<std::pair<Vector2D, int>> vector_costs;
+	vector<std::pair<Vector2D, float>> vector_costs;
 	float RandomFloat(float a, float b);
 	int RandomInt(int a, int b);
 	float Heuristic(Vector2D a, Vector2D b);
